tutorials/Streamlit/Diabete_prediction_app/requirements.txt
pandas==1.2.4
joblib==0.17.0
streamlit==0.87.0
Pillow==8.3.1
